Overhaul the typing system
- New system has a state machine approach where the GameEngine keeps
track of the state that the typing is in. It provides actions that then
can change the state of the game.

- State is now kept in subcomponents and made available to GameEngine
when it is needed, this I think makes it more manageable and
maintainable as parts are separated.

- I think it's a better approach, at least from what the mess that it
was before, even though it could be considered a bad way to do it.
Hopefully, in the future I learn an even better way.

- The actual typing now happens line by line which means less work as
everything else is cached.

- One big commit that in hindsight should have been split into smaller
commits, but things weren't working as expected and had to overhaul
twice so eh.