Add fixes to dialog.tsx
- Added props to enable fixes to be applied to dialog.

- Added a fix for input in dialog that is nested in a dropdown. This is
an issue, where when typing in input, moving the mouse selects the text
and causes you to overwrite the text.

- Added a fix for a controlled dialog nested in dropdown. This is an
issue because it was implemented badly but not sure how to fix it
properly.

- I think the trigger for the dialog should act as the
dropdown-menu-item component by using asChild to fix some of this stuff
without this roundabout approach, but I wasn't successful.